requirements.txt
numpy
opencv-python
mss
python-chess
chessimg2pos
stockfish
PyAutoGUI
human-mouse